buggy race between subscribers and api start, stop, update methods fixed, exceptions handled, stage 0 of bug fixing

+++ src/app/shared/utils/navigation/navigation.controller.ts
@@ -7,12 +7,11 @@ import {APIObject} from '../drawing/api.types';
 import * as d3 from 'd3';
 import {Subject} from 'rxjs/Subject';
 import {Observable} from 'rxjs/Observable';
+import {NavigationService} from './navigation.service';
+import {Geometry} from '../helper/geometry';
 import Metadata = APIObject.Metadata;
 import Path = APIObject.Path;
 import NavigationData = APIObject.NavigationData;
-import {NavigationService} from './navigation.service';
-import {Geometry} from '../helper/geometry';
-import {Subscription} from 'rxjs/Subscription';
 
 @Injectable()
 export class NavigationController {
@@ -27,7 +26,7 @@ export class NavigationController {
   private subscriptionDestructor: Subject<void> = new Subject<void>();
   private positionChanged: Subject<Point> = new Subject<Point>();
   private lines: Line[];
-  subscriptionPathService: Subscription;
+  private stoppedBeforeIsNavigationReady: boolean = false;
 
   constructor(
     private pathService: PathService,
@@ -37,19 +36,16 @@ export class NavigationController {
 
   handleNavigation(event: MessageEvent, floorId, container, scale) {
     const args: NavigationData = event.data.args.object;
-
-    console.log(args, 'is Navigation ready', this.isNavigationReady);
     if (this.isNavigationReady) {
       if (args.action === 'update') {
         this.updatePosition(args.position);
-      }
-      if (args.action === 'stop') {
-        this.stopNavigation();
       }
     } else if (args.action === 'start') {
       this.setNavigationPath(floorId, args.location, args.destination, args.accuracy, event, container, scale);
     } else if (args.action === 'update') {
       this.setLastCoordinates(args.position);
+    } else if (args.action === 'stop') {
+      this.stopNavigation();
     }
   }
 
@@ -61,22 +57,26 @@ export class NavigationController {
     this.container = container;
     this.destination = destination;
     this.scale = scale;
-
+    if (!this.subscriptionDestructor) {
+      this.subscriptionDestructor = new Subject<void>();
+    }
     if (this.isNavigationReady) {
       this.event.source.postMessage({type: 'navigation', action: 'working'}, this.event.origin);
       return;
     }
     this.event = event;
-    this.subscriptionPathService = this.pathService.getPathByFloorId(floorId).takeUntil(this.subscriptionDestructor).subscribe((lines: Line[]): void => {
+    this.pathService.getPathByFloorId(floorId).takeUntil(this.subscriptionDestructor).subscribe((lines: Line[]): void => {
       this.calculateNavigationPath(lines, location, destination, accuracy);
-      console.log('sdfsda sdafsd', this.isNavigationReady);
+      this.isNavigationReady = true;
+
       this.onPositionChanged().takeUntil(this.subscriptionDestructor).subscribe((pointUpdate: Point): void => {
         this.handlePathUpdate(pointUpdate);
       });
-      this.isNavigationReady = true;
-
       if (this.lastCoordinates) {
         this.updatePosition(this.lastCoordinates);
+      }
+      if (this.stoppedBeforeIsNavigationReady) {
+        this.stopNavigation();
       }
     });
   }
@@ -93,13 +93,17 @@ export class NavigationController {
     if (!!this.event) {
       this.event.source.postMessage({type: 'navigation', action: 'finished'}, this.event.origin);
     }
-    this.subscriptionPathService.unsubscribe();
-    this.isNavigationReady = false;
-    this.subscriptionDestructor.next();
-    this.subscriptionDestructor = null;
-    this.mapObjectService.removeObject(this.objectMetadata);
-    this.objectMetadata = null;
-    this.event = null;
+    if (this.isNavigationReady) {
+      this.isNavigationReady = false;
+      this.mapObjectService.removeObject(this.objectMetadata);
+      this.subscriptionDestructor.next();
+      this.subscriptionDestructor = null;
+      this.objectMetadata = null;
+      this.event = null;
+      this.stoppedBeforeIsNavigationReady = false;
+    } else {
+      this.stoppedBeforeIsNavigationReady = true;
+    }
   }
 
   private handlePathUpdate(pointUpdate: Point): void {
@@ -119,8 +123,7 @@ export class NavigationController {
     this.lines = this.lines.slice(findLineIndex);
     this.lines[0].startPoint = currentPointOnPath;
 
-    const points = this.createPointPathFromLinePath(this.scale, this.lines);
-    this.objectMetadata.object['points'] = points;
+    this.objectMetadata.object['points'] = this.createPointPathFromLinePath(this.scale, this.lines);
     this.redrawPath();
   }
 
@@ -143,22 +146,17 @@ export class NavigationController {
       type: 'POLYLINE'
     };
 
-    const points = this.createPointPathFromLinePath(this.scale, path);
-
-    this.objectMetadata.object['points'] = points;
+    this.objectMetadata.object['points'] = this.createPointPathFromLinePath(this.scale, path);
 
     this.objectMetadata.object = Object.assign((<Path>this.objectMetadata.object), {lines: path, color: '#906090', lineType: 'dotted'});
     this.redrawPath();
-  }
-
-  private calculatePointInCentimeters(scale: Scale, point: Point): Point {
-    return Geometry.calculatePointPositionInCentimeters(scale.getLenInPix(), scale.getRealDistanceInCentimeters(), point);
+    this.isNavigationReady = true;
   }
 
   private createPointPathFromLinePath(scale: Scale, path: Line[]): Point[] {
     return path.reduce((points, point) => {
-      points.push(this.calculatePointInCentimeters(scale, point.startPoint));
-      points.push(this.calculatePointInCentimeters(scale, point.endPoint));
+      points.push(Geometry.calculatePointInCentimeters(scale, point.startPoint));
+      points.push(Geometry.calculatePointInCentimeters(scale, point.endPoint));
       return points;
     }, []);
   }
